requirements.txt
python-telegram-bot==21.9
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
